Added some useful shell scripts to generate, simulate, and cleanup gromacs stuff.
I included the Lysozyme gro file and a sample mdp. First generate the binary and whatnot with command: generate.sh 1aki. Allow the simulation to run, and terminate it when ready with ctrl-c. Repeat simulations can be run with simulate.sh 1aki.
Generated files can be deleted with cleanup.sh.